Moje úpravy

+++ src/Flee.Net45/InternalTypes/Miscellaneous.cs
@@ -211,7 +211,7 @@ namespace Flee.InternalTypes
             // Our score is the average of the scores of each parameter.  The lower the score, the better the match.
             int sum = ComputeSum(parameters, argTypes);
 
-            return sum / argTypes.Length;
+            return sum / (float)argTypes.Length;
         }
 
         private static int ComputeSum(ParameterInfo[] parameters, Type[] argTypes)
